+++ assets/stylesheets/_components.scss
@@ -16,6 +16,7 @@
 @import 'blocks/comments/style';
 @import 'blocks/dismissible-card/style';
 @import 'blocks/domain-to-plan-nudge/style';
+@import 'blocks/like-button/style';
 @import 'blocks/plan-thank-you-card/style';
 @import 'blocks/post-edit-button/style';
 @import 'blocks/post-item/style';
@@ -84,7 +85,6 @@
 @import 'components/infinite-list/style';
 @import 'components/info-popover/style';
 @import 'components/input-chrono/style';
-@import 'components/like-button/style';
 @import 'components/logged-out-form/style';
 @import 'components/main/style';
 @import 'components/marketing-survey/cancel-purchase-form/style';
