+++ client/blocks/like-button/icons.jsx
