+++ client/blocks/like-button/index.jsx
