+++ client/reader/comments/comment-likes.jsx
@@ -6,7 +6,7 @@ import { bindActionCreators } from 'redux';
 import { translate } from 'i18n-calypso';
 
 // Internal dependencies
-import LikeButton from 'components/like-button/button';
+import LikeButton from 'blocks/like-button/button';
 import {
 	recordAction,
 	recordGaEvent,
